修复bug
